chore(Controllers): Added Cache Controller
